backend/requirements.txt
Django>=4.1.3,<4.2
djangorestframework>=3.14.0,<3.15
drf-spectacular>=0.24.2,<0.25
psycopg2>=2.9.5,<3.0
django-filter>=22.1,<22.2
django-cors-headers>=3.13,<3.14
PyJWT>=2.6.0,<2.7
djangorestframework-simplejwt>=5.2.2,<5.3
dj-rest-auth>=2.2.5,<2.3
django-allauth>=0.51.0,<0.52